rename ProductCardList component to Products
